Rest api -> master

+++ src/main/java/com/netcracker/skillstable/repos/EntTypeAttrRepo.java
@@ -2,16 +2,14 @@ package com.netcracker.skillstable.repos;
 
 import com.netcracker.skillstable.model.EntTypeAttr;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 import java.util.List;
 
 @RepositoryRestResource(collectionResourceRel = "entTypeAttr", path = "entTypeAttr")
-public interface EntTypeAttrRepo extends JpaRepository<EntTypeAttr, Long> {
-    List<EntTypeAttr> findByEntityTypeId(Long entTypeId);
+public interface EntTypeAttrRepo extends JpaRepository<EntTypeAttr, Integer> {
+    List<EntTypeAttr> findByEntityTypeId(Integer entTypeId);
 
-    EntTypeAttr findByEntityTypeIdAndAttributeId(Long entTypeId, Long attrId);
+    EntTypeAttr findByEntityTypeIdAndAttributeId(Integer entTypeId, Integer attrId);
 }
